Implement Spawning Enemy Waves
Added BeginPlay() to set CurrentWave to 0 at the start.
Created an AActor* array ActiveEnemies to hold enemies that are spawned for the current round.
Added an int LastSpawnLocationElement for use in spawning actors.
Added void SpawnWave() which takes all the enemies in the current wave and spawns them randomly around the given spawn points.